Shortest distance between adjacent numbers
The file contains the program to find the shortest distance between adjacent numbers. 
Here, two numbers in the array are adjacent, iff there exists no values in the range of two numbers in the array.
It is also necessary that the two numbers are different. 
The distance between these adjacent numbers are the difference in the indices of the numbers in the array. 
The program outputs the minimum distance between the many adjacent numbers and -1 if there exists no adjacent numbers.